Add path for Task components
