+++ .gitignore
@@ -31,4 +31,4 @@ student-database-2aa8d-5888003c045e.json
 Standards K-8.xlsx
 testAnswerKey.csv
 
-firebase-export-1707978959241OUHu3k
+firebase-export-*
